+++ Runtime/Components/Movements/Movement.cs
@@ -6,23 +6,22 @@ namespace DapperDino.DapperTools.Components.Movements
     public class Movement
     {
         private readonly CharacterController controller;
-        private readonly List<IMovementModifier> modifiers;
+        private readonly List<IMovementModifier> modifiers = new List<IMovementModifier>();
 
-        public Movement(CharacterController controller)
-        {
-            this.controller = controller;
-            modifiers = new List<IMovementModifier>();
-        }
+        public Movement(CharacterController controller) => this.controller = controller;
 
         public void Tick(float deltaTime)
         {
-            Vector3 movement = Vector3.zero;
+            // Initiate this frame's movement to be zero (0, 0, 0)
+            var movement = Vector3.zero;
 
+            // Add on each modifier's movement value
             foreach (var modifier in modifiers)
             {
                 movement += modifier.Value;
             }
 
+            // Move the character controller by the calculated movement
             controller.Move(movement * deltaTime);
         }